feat: separate Wrapper and Stack into independent classes
